+++ src/lib.rs
@@ -48,12 +48,15 @@ extern crate log;
 extern crate mongodb;
 extern crate serde;
 
-pub use bson::Document; // NOTE: TODO: no real need to `pub use` this. Remove in `1.0`.
-pub use mongodb::coll::options::FindOptions; // NOTE: TODO: no real need to `pub use` this. Remove in `1.0`.
-
 pub mod migration;
 pub mod model;
 
 // Expose lower symbols in the top level module.
-pub use model::Model;
-pub use model::basic_index_options;
+pub use migration::{
+    IntervalMigration,
+    Migration,
+};
+pub use model::{
+    basic_index_options,
+    Model,
+};