+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 changelog
 =========
 
+## 0.5
+- a migrations system has been added, closing [#3](https://github.com/thedodd/wither/issues/3) & [#4](https://github.com/thedodd/wither/issues/4). The important part of this feature set is the `IntervalMigration` type.
+- **Backwards Incompatibility:** `Model::sync` has received some updates. It now synchronizes a model's indexes as well as its migrations. It no longer panics. It will now return a `Result`, offering users a greater level of control on behavior.
+- reexports of `bson::Document` & `mongodb::coll::options::FindOptions` have been removed. They must be vestiges of my early development on this crate.
+- now testing all builds against MongoDB 3.2.x & 3.4.x. Will add 3.6.x when it is available on hub.docker.com.
+
 ## 0.4
 - adds `Model.update`.